% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_figure_utils.R
\name{include_fig}
\alias{include_fig}
\title{FUNCTION_TITLE}
\usage{
include_fig(FIGPATH = "path_01.pdf", WIDTH = 0.8, PAGE = 1)
}
\arguments{
\item{FIGPATH}{PARAM_DESCRIPTION, Default: 'path_01.pdf'}

\item{WIDTH}{PARAM_DESCRIPTION, Default: 0.8}

\item{PAGE}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
}
