% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_figure_utils.R
\name{sub_fig}
\alias{sub_fig}
\title{FUNCTION_TITLE}
\usage{
sub_fig(SUBPLACE = "H", SUBTEXTWIDTH = 0.7, SUBFIGWIDTH = 1,
  SUBPAGE = 1, SUBFIGPATH = "path_01.pdf",
  SUBCAPTION = "subcaption_01", SUBLABEL = "fig:subfig_01",
  SUBFIGSPACE = "\\\\quad")
}
\arguments{
\item{SUBPLACE}{PARAM_DESCRIPTION, Default: 'H'}

\item{SUBTEXTWIDTH}{PARAM_DESCRIPTION, Default: 0.7}

\item{SUBFIGWIDTH}{PARAM_DESCRIPTION, Default: 1}

\item{SUBPAGE}{PARAM_DESCRIPTION, Default: 1}

\item{SUBFIGPATH}{PARAM_DESCRIPTION, Default: 'path_01.pdf'}

\item{SUBCAPTION}{PARAM_DESCRIPTION, Default: 'subcaption_01'}

\item{SUBLABEL}{PARAM_DESCRIPTION, Default: 'fig:subfig_01'}

\item{SUBFIGSPACE}{PARAM_DESCRIPTION, Default: 'quad'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
}
