% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{tex_figures}
\alias{tex_figures}
\title{FUNCTION_TITLE}
\usage{
tex_figures(captions, figure_files_dir = "figure",
  analysis = "efficacy", only_tex = TRUE)
}
\arguments{
\item{captions}{character, Named character vector. Names are the same as the figure filenames in deliv/table and the value is the caption}

\item{figure_files_dir}{character, Path to location of the analysis figures directory in deliv, Default: 'figure'}

\item{analysis}{character, Name of the analysis to place in the tex comment block, Default: 'efficacy'}

\item{only_tex}{logical, return tibble or only the tex as character vector, Default: TRUE}
}
\value{
character/tibble depending on only_tex
}
\description{
FUNCTION_DESCRIPTION
}
