% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_table_utils.R
\name{tex_include}
\alias{tex_include}
\title{Generate an include call for tex}
\usage{
tex_include(path)
}
\arguments{
\item{path}{character, path to file}
}
\value{
character
}
\description{
wraps include around a path for tex
}
