% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex.R
\name{tex_tables}
\alias{tex_tables}
\title{FUNCTION_TITLE}
\usage{
tex_tables(captions, table_files_dir = "table", analysis = "efficacy",
  only_tex = TRUE)
}
\arguments{
\item{captions}{character, Named character vector. Names are the same as the figure filenames in deliv/table and the value is the caption}

\item{table_files_dir}{character, path to location of the analysis tables directory in deliv, Default: 'table'}

\item{analysis}{character, Name of the analysis to place in the tex comment block, Default: 'efficacy'}

\item{only_tex}{logical, return tibble or only the tex as character vector, Default: TRUE}
}
\value{
character/tibble depending on only_tex

charater
}
\description{
FUNCTION_DESCRIPTION
}
