% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_update.R
\name{tex_update}
\alias{tex_update}
\title{Update tex file with table/figure enviornments}
\usage{
tex_update(tex, file, section = "Efficacy", type = "Tables",
  show_only_new = FALSE, overwrite = FALSE)
}
\arguments{
\item{tex}{character, tex lines to edit in file}

\item{file}{character, Path of file to edit}

\item{section}{character, Name of section tex lines that will be part of the tex comment block, Default: 'Efficacy'}

\item{type}{character, Type of output (Tables or Figures) that will be part of the tex comment block, Default: 'Tables'}

\item{show_only_new}{logical, return to the console what only new section (useful for testing), Default: FALSE}

\item{overwrite}{logical, overwrite the file, Default: FALSE}
}
\value{
character
}
\description{
Use script to update figures.tex and tables.tex files in a centralized workflow
}
