% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_figure_utils.R
\name{wrap_fig}
\alias{wrap_fig}
\title{FUNCTION_TITLE}
\usage{
wrap_fig(SUBFIGS = include_fig(), PLACE = "H",
  CAPTION = "caption_01", LABEL = "fig:fig01")
}
\arguments{
\item{SUBFIGS}{PARAM_DESCRIPTION, Default: include_fig()}

\item{PLACE}{PARAM_DESCRIPTION, Default: 'H'}

\item{CAPTION}{PARAM_DESCRIPTION, Default: 'caption_01'}

\item{LABEL}{PARAM_DESCRIPTION, Default: 'fig:fig01'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
}
