% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_table_utils.R
\name{wrap_tbl}
\alias{wrap_tbl}
\title{FUNCTION_TITLE}
\usage{
wrap_tbl(TABLE, CAPTION = "caption_01", LABEL = "tbl:tbl01",
  PLACE = "!htbp")
}
\arguments{
\item{TABLE}{PARAM_DESCRIPTION}

\item{CAPTION}{PARAM_DESCRIPTION, Default: 'caption_01'}

\item{LABEL}{PARAM_DESCRIPTION, Default: 'tbl:tbl01'}

\item{PLACE}{PARAM_DESCRIPTION, Default: '!htbp'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
}
